Move getZodType & getZodSchema to separate file
